Moved analysis to util and testsystems to tests
