+++ bhmm/plots.py
@@ -49,9 +49,8 @@ def plot_state_assignments(model, s_t, o_t, tau=1.0, time_units=None, obs_label=
     Example
     -------
 
-    >>> import tempfile
+    >>> from bhmm.tests import testsystems    >>> import tempfile
     >>> filename = tempfile.NamedTemporaryFile().name
-    >>> from bhmm import testsystems
     >>> [model, O, S, bhmm_sampler] = testsystems.generate_random_bhmm(nstates=3, ntrajectories=1, length=10000)
     >>> plot_state_assignments(model, S[0], O[0], pdf_filename=filename)
