+++ bhmm/util/analysis.py
