+++ examples/rnase-h-d10a/generate-figure.py
@@ -4,7 +4,7 @@ Generate plots for synthetic three-state force spectroscopy model.
 
 """
 
-from bhmm import testsystems
+from bhmm.tests import testsystems
 from bhmm import BHMM
 from bhmm import MaximumLikelihoodEstimator
 from bhmm import plots
